backend2/requirements.txt
fastapi==0.104.1
uvicorn==0.22.0
sqlalchemy==2.0.15
pydantic==2.3.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
alembic==1.11.1
pymysql==1.0.3
openpyxl